Flint Michigan Lead Data

Lead poisoning can be a serious problem associated with drinking tap water. Many older water pipes are made of lead. Over time, the pipes corrode, releasing lead into the drinking water. In April 2014, the city of Flint Michigan switched its water supply to the Flint River in an effort to save money. The Michigan Department of Environmental Quality (MDEQ) tested the water at the time and declared it safe to drink. Officials were supposed to test at least 100 homes, targeting those most at risk. The U.S. Environmental Protection Agency (EPA)’s Lead and Copper Rule states that if lead concentrations exceed an action level of 15 parts per billion (ppb) in more than 10% of homes sampled, then actions must be undertaken to control corrosion, and the public must be informed.

Usage

FlintMDEQ

Format

flint

A data frame with 71 rows and 1 column:

lead

Lead concentration per household, measured in parts per billion.