ISCAM: Companion to the Book "Investigating Statistical Concepts, Applications, and Methods"

Introductory statistics methods to accompany "Investigating Statistical Concepts, Applications, and Methods" (ISCAM) by Beth Chance & Allan Rossman. Tools to introduce statistical concepts with a focus on simulation approaches. Functions are verbose, designed to provide ample output for students to understand what each function does. Additionally, most functions are accompanied with plots. The package is designed to be used in an educational setting alongside the ISCAM textbook.

See also

Author

Maintainer: Beth Chance bchance@calpoly.edu [copyright holder]

Authors:

Visruth Srimath Kandali visruth@gmail.com (ORCID)